Fx provisioner test and add main file tests
